phase_3/backend/requirements.txt
fastapi==0.124.0
uvicorn==0.34.0
sqlmodel==0.0.22
pydantic==2.11.7
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
alembic==1.13.1
python-dotenv==1.0.1
pytest==7.4.3
httpx==0.28.1
mangum==0.17.0
google-generativeai==0.8.6
fastapi-socketio==0.0.10
python-socketio==5.16.0
python-engineio==4.13.0
bidict==0.23.1
simple-websocket==1.1.0
wsproto==1.3.2